Fixed Available Flights bug
Ongoing fix for payment details
